% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{catchplot.msam}
\alias{catchplot.msam}
\title{Catch plot}
\usage{
\method{catchplot}{msam}(fit, obs.show = TRUE, drop = 0, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
Catch plot
}
\author{
Christoffer Moesgaard Albertsen
}
