% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catchtable.msam}
\alias{catchtable.msam}
\title{Catch Table}
\usage{
\method{catchtable}{msam}(fit, obs.show = FALSE, returnList = FALSE,
  ...)
}
\arguments{
\item{fit}{msam object}

\item{obs.show}{should observed catches be included?}

\item{returnList}{If true, a list of matrices is returned}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Catch Table
}
\details{
Table of predicted catch in weight
}
\author{
Christoffer Moesgaard Albertsen
}
