% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.msam}
\alias{coef.msam}
\title{Extract parameter estimates of msam object}
\usage{
\method{coef}{msam}(object, ...)
}
\arguments{
\item{object}{msam object}

\item{...}{extra arguments not used}
}
\value{
a vector of parameter estimates
}
\description{
Extract parameter estimates of msam object
}
\author{
Christoffer Moesgaard Albertsen
}
