% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplot.R
\name{corplot.msam}
\alias{corplot.msam}
\title{Correlation and partial correlation plot between survival processes}
\usage{
\method{corplot}{msam}(object, ...)
}
\arguments{
\item{object}{msam object}

\item{...}{Other parameters not currently used}
}
\value{
A matrix of correlations (lower triangular) and partial correlations (upper triangular)
}
\description{
Correlation and partial correlation plot between survival processes
}
\author{
Christoffer Moesgaard Albertsen
}
