% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicReferencepoints.R
\name{deterministicReferencepoints.msam}
\alias{deterministicReferencepoints.msam}
\title{Deterministic reference points}
\usage{
\method{deterministicReferencepoints}{msam}(
  fit,
  referencepoints,
  catchType = "catch",
  nYears = 300,
  Fsequence = seq(0, 2, len = 50),
  aveYears = lapply(fit, function(x) max(x$data$years) + (-9:0)),
  selYears = lapply(fit, function(x) max(x$data$years)),
  biasCorrect = FALSE,
  newton.control = list(),
  ...
)
}
\arguments{
\item{fit}{}

\item{referencepoints}{}

\item{catchType}{}

\item{nYears}{}

\item{Fsequence}{}

\item{aveYears}{}

\item{selYears}{}

\item{biasCorrect}{}

\item{newton.control}{}

\item{...}{}
}
\value{
results
}
\description{
Deterministic reference points
}
\author{
Christoffer Moesgaard Albertsen
}
