% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{erbplot.msam}
\alias{erbplot.msam}
\title{ERB plot}
\usage{
\method{erbplot}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
ERB plot
}
\author{
Christoffer Moesgaard Albertsen
}
