% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{erbtable.msam}
\alias{erbtable.msam}
\title{SSB Table}
\usage{
\method{erbtable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Effective reproductive biomass Table
}
\details{
Table of estimated effective reproductive biomass
}
\author{
Christoffer Moesgaard Albertsen
}
