% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{faytable.msam}
\alias{faytable.msam}
\title{F-at-age Table}
\usage{
\method{faytable}{msam}(fit, returnList = FALSE, ...)
}
\arguments{
\item{fit}{msam object}

\item{returnList}{If true, a list of matrices is returned}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
F-at-age table
}
\details{
Table of estimated age-wise fishing mortality
}
\author{
Christoffer Moesgaard Albertsen
}
