% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fbarplot.msam}
\alias{fbarplot.msam}
\title{Fbar plot for msam object}
\usage{
\method{fbarplot}{msam}(fit, partial = FALSE, drop = 0, page = NULL, ...)
}
\arguments{
\item{fit}{}

\item{...}{}
}
\description{
Fbar plot for msam object
}
\author{
Christoffer Moesgaard Albertsen
}
