% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fbartable.msam}
\alias{fbartable.msam}
\title{Fbar Table}
\usage{
\method{fbartable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Fbar Table
}
\details{
Table of estimated average fishing mortalities
}
\author{
Christoffer Moesgaard Albertsen
}
