% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fselectivityplot.msam}
\alias{fselectivityplot.msam}
\title{F selectivity plot}
\usage{
\method{fselectivityplot}{msam}(
  fit,
  cexAge = 1,
  type = c("bar", "line"),
  col = .plotcols.crp(length(fit)),
  ...
)
}
\arguments{
\item{fit}{msam object}

\item{cexAge}{size of Age label}

\item{type}{bar for separate bar plots, line for overlayed line plots}

\item{col}{Line colors for line plot}

\item{...}{extra arguments for plotting}
}
\description{
F selectivity plot
}
\author{
Christoffer Moesgaard Albertsen
}
