% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.R
\name{gen2PCA}
\alias{gen2PCA}
\title{Prepare allele matrix for PCA analysis}
\usage{
gen2PCA(x, alleleMeans)
}
\arguments{
\item{x}{output from \\link{read.gen}}

\item{alleleMeans}{Mean allele frequencies to impute. Is calculated if missing.}
}
\value{
allele data for PCA analysis
}
\description{
Prepare allele matrix for PCA analysis
}
\author{
Christoffer Moesgaard Albertsen
}
