% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisam.R
\name{getAllDerivedValues.msam}
\alias{getAllDerivedValues.msam}
\title{Update multi-stock sam fit with additional derived values}
\usage{
\method{getAllDerivedValues}{msam}(fit)
}
\arguments{
\item{fit}{multi-stock sam fit returned by multisam.fit}
}
\value{
Updated sam fit
}
\description{
Update multi-stock sam fit with additional derived values
}
