% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencepoints.R, R/sam_jacobian.R
\name{grad}
\alias{grad}
\title{Calculate gradient of a function}
\usage{
grad(
  func,
  x,
  h = abs(1e-04 * x) + 1e-04 * (abs(x) < sqrt(.Machine$double.eps/7e-07)),
  ...
)

grad(
  func,
  x,
  h = abs(1e-04 * x) + 1e-04 * (abs(x) < sqrt(.Machine$double.eps/7e-07)),
  ...
)
}
\arguments{
\item{func}{function}

\item{x}{parameter values}

\item{h}{step size}

\item{...}{passed to func}
}
\value{
gradient vector

gradient vector
}
\description{
Calculate gradient of a function

Calculate gradient of a function
}
\author{
Christoffer Moesgaard Albertsen
}
