% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{holdout}
\alias{holdout}
\alias{holdout.msam}
\title{Holdout forecast for validation}
\usage{
holdout(fit, nYears, ...)

\method{holdout}{msam}(fit, nYears, forecastYears = 1, ncores = 1, ...)
}
\arguments{
\item{fit}{msam fit}

\item{nYears}{Number of years to hold out}

\item{...}{Arguments passed to modelforecast}
}
\value{
A model forecast
}
\description{
Holdout forecast for validation
}
\author{
Christoffer Moesgaard Albertsen
}
