% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lifeexpectancyplot.msam}
\alias{lifeexpectancyplot.msam}
\title{Life expectancy plot}
\usage{
\method{lifeexpectancyplot}{msam}(
  fit,
  atRecruit = TRUE,
  col = .plotcols.crp(length(fit)),
  ylimAdd = max(sapply(fit, function(x) x$conf$maxAge)),
  ...
)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
Life expectancy plot
}
\author{
Christoffer Moesgaard Albertsen
}
