% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{lifeexpectancytable.msam}
\alias{lifeexpectancytable.msam}
\title{SSB Table}
\usage{
\method{lifeexpectancytable}{msam}(fit, atRecruit = TRUE, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Life expectancy Table
}
\details{
Table of estimated life expectancy.
}
\author{
Christoffer Moesgaard Albertsen
}
