% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{modeltable.msamset}
\alias{modeltable.msamset}
\title{Model table}
\usage{
\method{modeltable}{msamset}(fits, ...)
}
\arguments{
\item{fits}{msamset object}

\item{...}{extra arguments}
}
\value{
A matrix of model information
}
\description{
Model table
}
\author{
Christoffer Moesgaard Albertsen
}
