% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ntable.msam}
\alias{ntable.msam}
\title{N Table}
\usage{
\method{ntable}{msam}(fit, returnList = FALSE, ...)
}
\arguments{
\item{fit}{msam object}

\item{returnList}{If true, a list of matrices is returned}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
N Table
}
\details{
Table of estimated numbers-at-age
}
\author{
Christoffer Moesgaard Albertsen
}
