% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parplot.msam}
\alias{parplot.msam}
\title{Parameter plot}
\usage{
\method{parplot}{msam}(fit, cor.report.limit = 0.95, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
Parameter plot
}
\author{
Christoffer Moesgaard Albertsen
}
