% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parplot.msamset}
\alias{parplot.msamset}
\title{Parameter plot}
\usage{
\method{parplot}{msamset}(fit, cor.report.limit = 0.95, ...)
}
\arguments{
\item{fit}{msam object}

\item{cor.report.limit}{Not used}

\item{...}{extra arguments for plotting}
}
\description{
Parameter plot
}
\author{
Christoffer Moesgaard Albertsen
}
