% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{partable.msam}
\alias{partable.msam}
\title{Parameter table}
\usage{
\method{partable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Parameter table
}
\details{
Table of estimated parameters
}
\author{
Christoffer Moesgaard Albertsen
}
