% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotit.msam}
\alias{plotit.msam}
\title{Function to actually do the plotting}
\usage{
\method{plotit}{msam}(fit, what, x = lapply(attr(fit, "m_data")$sam,
  function(x) x$years), ylab = what, xlab = "Years", ex = numeric(0),
  trans = function(x) x, add = FALSE, ci = TRUE, cicol = gray(0.5, alpha
  = 0.5), addCI = NA, drop = 0, unnamed.basename = "current",
  xlim = NULL, ciAlpha = 0.3, col = .plotcols.crp(length(fit)),
  extraLabel = NULL, ...)
}
\description{
Function to actually do the plotting
}
