% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.msam}
\alias{print.msam}
\title{Print result from multisam.fit}
\usage{
\method{print}{msam}(x, ...)
}
\arguments{
\item{x}{A msam object}

\item{...}{Other parameters passes to logLik}
}
\description{
Print output from multisam.fit
}
