% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{print.msamcoef}
\alias{print.msamcoef}
\title{Print msam coef object}
\usage{
\method{print}{msamcoef}(x, ...)
}
\arguments{
\item{x}{msamcoef object}

\item{...}{Not used}
}
\description{
Print msam coef object
}
\author{
Christoffer Moesgaard Albertsen
}
