% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gen.R
\name{read.gen}
\alias{read.gen}
\title{Read genepop data files}
\usage{
read.gen(
  f,
  pop.names,
  sort.loci = FALSE,
  sort.individuals = FALSE,
  NAlleleKeep = NA
)
}
\arguments{
\item{f}{file name}

\item{pop.names}{population names. If missing, the ID of the last individual is used}

\item{sort.loci}{Should loci be sorted by names?}

\item{sort.individuals}{Should individuals be sorted by id?}
}
\value{
an allele array
}
\description{
Read genepop data files
}
\author{
Christoffer Moesgaard Albertsen
}
