% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{recplot.msam}
\alias{recplot.msam}
\title{Recruitment plot}
\usage{
\method{recplot}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
Recruitment plot
}
\author{
Christoffer Moesgaard Albertsen
}
