% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rectable.msam}
\alias{rectable.msam}
\title{Recruitment Table}
\usage{
\method{rectable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
Recruitment Table
}
\details{
Table of estimated recruitment to the fisheries.
}
\author{
Christoffer Moesgaard Albertsen
}
