% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{srplot.msam}
\alias{srplot.msam}
\title{Stock-recruitment plot}
\usage{
\method{srplot}{msam}(
  fit,
  stock,
  textcol = "red",
  linetype = "l",
  linecol = "black",
  polycol = sapply(.plotcols.crp(length(fit)), function(cc) do.call("rgb",
    c(as.list(col2rgb(cc)[, 1]), list(alpha = 0.1)))),
  polyborder = sapply(.plotcols.crp(length(fit)), function(cc) do.call("rgb",
    c(as.list(col2rgb(cc)[, 1]), list(alpha = 0.3)))),
  polylty = 3,
  polylwd = 1,
  add = FALSE,
  CIlevel = 0.95,
  addCurve = TRUE,
  col = .plotcols.crp(length(fit)),
  ...
)
}
\arguments{
\item{fit}{msam object}

\item{textcol}{Text color}

\item{add}{Should the figure be added to a current plot?}

\item{col}{Line colors}

\item{...}{extra arguments for plotting}
}
\description{
Stock-recruitment plot
}
\author{
Christoffer Moesgaard Albertsen
}
