% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ssbplot.msam}
\alias{ssbplot.msam}
\title{SSB plot}
\usage{
\method{ssbplot}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
SSB plot
}
\author{
Christoffer Moesgaard Albertsen
}
