% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ssbtable.msam}
\alias{ssbtable.msam}
\title{SSB Table}
\usage{
\method{ssbtable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
SSB Table
}
\details{
Table of estimated spawning stock biomasses.
}
\author{
Christoffer Moesgaard Albertsen
}
