% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_covstructure.R
\name{suggestCorStructure}
\alias{suggestCorStructure}
\title{Suggest correlation structure}
\usage{
suggestCorStructure(
  x,
  nAgeClose = 1,
  noConnection = FALSE,
  onlyCloseAge = TRUE,
  noCorInArea = TRUE,
  onlyCorInArea = FALSE,
  noCorBetween = matrix(0, 0, 2)
)
}
\arguments{
\item{x}{a samset}

\item{nAgeClose}{Number of close ages to correlate}

\item{noConnection}{If TRUE, no correlations}

\item{onlyCloseAge}{If TRUE, age classes no less than nAgeClose are un-correlated}

\item{noCorInArea}{If TRUE, age classes in same area/stock are un-correlated}

\item{onlyCorInArea}{If TRUE, only age classes in same area/stock are correlated}

\item{noCorBetween}{matrix of area/stock numbers that should not be correlated}
}
\value{
A boolean matrix
}
\description{
Construct valid (band) correlation structures for multisam.fit
}
\author{
Christoffer Moesgaard Albertsen
}
