% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.msam}
\alias{summary.msam}
\title{Summary of msam object}
\usage{
\method{summary}{msam}(object, returnList = FALSE, digits = c(0, 0, 3), ...)
}
\arguments{
\item{object}{msam object}

\item{returnList}{If true, a list of table is returned}

\item{digits}{digits for rounding output. Vector of length 3 corresponding to recruitment, ssb, and fbar.}

\item{...}{not used}
}
\value{
a summary table
}
\description{
Summary of msam object
}
\author{
Christoffer Moesgaard Albertsen
}
