% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableit}
\alias{tableit}
\alias{tableit.msam}
\title{Table helper}
\usage{
tableit(fit, what, x = lapply(attr(fit, "m_data")$sam, function(x)
  x$years), trans = function(x) x, returnList = FALSE, ...)

\method{tableit}{msam}(fit, what, x = lapply(attr(fit, "m_data")$sam,
  function(x) x$years), trans = function(x) x, returnList = FALSE, ...)
}
\arguments{
\item{fit}{msam object}

\item{what}{quoted name of what to extract}

\item{x}{rownames of table}

\item{trans}{function to be applied}

\item{returnList}{If true, a list of matrices is returned}

\item{...}{extra arguments not used}
}
\value{
matrix of estimates and confidence intervals
}
\description{
Table helper
}
\author{
Christoffer Moesgaard Albertsen
}
