% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsbplot.msam}
\alias{tsbplot.msam}
\title{TSB plot}
\usage{
\method{tsbplot}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
TSB plot
}
\author{
Christoffer Moesgaard Albertsen
}
