% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tsbtable.msam}
\alias{tsbtable.msam}
\title{TSB Table}
\usage{
\method{tsbtable}{msam}(fit, ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{Other parameters}
}
\value{
A matrix of estimates and confidence intervals
}
\description{
TSB Table
}
\details{
Table of estimated total stock biomasses.
}
\author{
Christoffer Moesgaard Albertsen
}
