% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{yearslostplot.msam}
\alias{yearslostplot.msam}
\title{Life Years lost to fishing plot}
\usage{
\method{yearslostplot}{msam}(fit, cause = c("Fishing", "Other", "LifeExpectancy"), ...)
}
\arguments{
\item{fit}{msam object}

\item{...}{extra arguments for plotting}
}
\description{
Life Years lost to fishing plot
}
\author{
Christoffer Moesgaard Albertsen
}
