% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ypr.msam}
\alias{ypr.msam}
\title{Yield per recruit calculation}
\usage{
\method{ypr}{msam}(fit, Flimit = 2, Fdelta = 0.01,
  aveYears = lapply(attr(fit, "m_data")$sam, function(x) min(15,
  length(x$years))), ageLimit = 100, ...)
}
\arguments{
\item{fit}{msam object}

\item{Flimit}{Upper limit for Fbar}

\item{Fdelta}{increments on the Fbar axis}

\item{aveYears}{A list/vector of same length as the number of stocks with number of years to average over}

\item{ageLimit}{Oldest age used}

\item{...}{not used}
}
\value{
A list of samypr objects
}
\description{
Yield per recruit calculation
}
\author{
Christoffer Moesgaard Albertsen
}
