% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{addTrans}
\alias{addTrans}
\title{Add transparency to a color}
\usage{
addTrans(name, alpha = 1)
}
\arguments{
\item{name}{Name of color}

\item{alpha}{Alpha level}
}
\value{
A hex color string
}
\author{
Christoffer Moesgaard Albertsen
}
