% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicorplot.R
\name{corplot}
\alias{corplot}
\title{Correlation plot for assessment model}
\usage{
corplot(object, ...)
}
\arguments{
\item{object}{Assessment object}

\item{...}{Other arguments}
}
\value{
It depends...
}
\author{
Christoffer Moesgaard Albertsen
}
