% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fitplot.msam}
\alias{fitplot.msam}
\title{Fit plot}
\usage{
\method{fitplot}{msam}(fit, stock, log = TRUE,
  fleets = lapply(attr(fit, "m_data")$sam, function(x) unique(x$aux[,
  "fleet"])), ...)
}
\arguments{
\item{fit}{msam object}

\item{stock}{Stock to plot for}

\item{log}{Plot on log-scale?}

\item{fleets}{Fleets to plot}

\item{...}{extra arguments for plotting}
}
\description{
Fit plot
}
\author{
Christoffer Moesgaard Albertsen
}
