% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_jacobian.R
\name{hessian}
\alias{hessian}
\title{Calculate hessian of a function}
\usage{
hessian(
  func,
  x,
  h = abs(1e-04 * x) + 1e-04 * (abs(x) < sqrt(.Machine$double.eps/7e-07)),
  columns = seq_along(x),
  ...
)
}
\arguments{
\item{func}{function}

\item{x}{parameter values}

\item{h}{step size}

\item{...}{passed to func}
}
\value{
jacobian matrix
}
\description{
Calculate hessian of a function
}
\note{
Could be made more accurate in some cases
}
