% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{modelforecast.msam}
\alias{modelforecast.msam}
\title{Forecast of a multiStockassessment}
\usage{
\method{modelforecast}{msam}(
  fit,
  constraints = NULL,
  fscale = NULL,
  catchval = NULL,
  fval = NULL,
  nextssb = NULL,
  landval = NULL,
  nosim = 1000,
  year.base = unlist(min(sapply(fit, function(x) max(x$data$years)))),
  ave.years = lapply(fit, function(x) max(x$data$years) + (-9:0)),
  rec.years = lapply(fit, function(x) numeric(0)),
  rec.fixval = lapply(fit, function(x) numeric(0)),
  rec.fixvalVar = lapply(fit, function(x) (1e-06)^2),
  label = NULL,
  overwriteSelYears = NULL,
  deterministicF = FALSE,
  processNoiseF = FALSE,
  fixedFdeviation = FALSE,
  useFHessian = FALSE,
  resampleFirst = !is.null(nosim) && nosim > 0,
  resampleParameters = FALSE,
  useModelLastN = TRUE,
  fixFirstN = FALSE,
  customSel = NULL,
  lagR = FALSE,
  splitLD = FALSE,
  addTSB = FALSE,
  biasCorrect = FALSE,
  returnAllYears = FALSE,
  returnObj = FALSE,
  progress = nosim > 0 && ncores == 1,
  estimate = median,
  silent = TRUE,
  newton_config = NULL,
  custom_pl = NULL,
  useNonLinearityCorrection = (nosim > 0 && !deterministicF),
  ncores = 1,
  overwriteBioProcessModel = FALSE,
  useManagementLag = FALSE,
  assessmentErrorMean_F = 0,
  assessmentErrorRho_F = 0,
  assessmentErrorSigma_F = 0,
  assessmentErrorMean_N = 0,
  assessmentErrorRho_N = 0,
  assessmentErrorSigma_N = 0,
  assessmentErrorMean_M = 0,
  assessmentErrorRho_M = 0,
  assessmentErrorSigma_M = 0,
  assessmentErrorMean_Mat = 0,
  assessmentErrorRho_Mat = 0,
  assessmentErrorSigma_Mat = 0,
  assessmentErrorMean_SW = 0,
  assessmentErrorRho_SW = 0,
  assessmentErrorSigma_SW = 0,
  assessmentErrorMean_CW = 0,
  assessmentErrorRho_CW = 0,
  assessmentErrorSigma_CW = 0,
  implementationErrorRho_F = 0,
  ...
)
}
\arguments{
\item{fit}{...}

\item{fscale}{...}

\item{catchval}{...}

\item{fval}{...}

\item{nextssb}{...}

\item{landval}{...}

\item{nosim}{...}

\item{year.base}{...}

\item{ave.years}{...}

\item{rec.years}{...}

\item{label}{...}

\item{overwriteSelYears}{...}

\item{deterministicF}{...}

\item{processNoiseF}{...}

\item{customSel}{...}

\item{lagR}{...}

\item{splitLD}{...}

\item{addTSB}{...}

\item{biasCorrect}{...}

\item{returnAllYears}{...}

\item{...}{...}
}
\value{
msamforecast
}
\description{
Forecast of a multiStockassessment
}
