% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{modelforecast.msam}
\alias{modelforecast.msam}
\title{Forecast of a multiStockassessment}
\usage{
\method{modelforecast}{msam}(
  fit,
  constraints = NULL,
  fscale = NULL,
  catchval = NULL,
  fval = NULL,
  nextssb = NULL,
  landval = NULL,
  findMSY = NULL,
  hcr = NULL,
  nosim = 0,
  year.base = sapply(fit, function(x) max(x$data$years)),
  ave.years = lapply(fit, function(x) numeric(0)),
  rec.years = lapply(fit, function(x) numeric(0)),
  label = NULL,
  overwriteSelYears = NULL,
  deterministicF = FALSE,
  processNoiseF = TRUE,
  resampleFirst = !is.null(nosim) && nosim > 0,
  customSel = NULL,
  lagR = FALSE,
  splitLD = FALSE,
  addTSB = FALSE,
  biasCorrect = FALSE,
  returnAllYears = FALSE,
  useUniroot = FALSE,
  nCatchAverageYears = rep(1, length(fit)),
  returnObj = FALSE,
  hcrConf = lapply(fit, function(x) numeric(0)),
  hcrCurrentSSB = 0,
  progress = TRUE,
  estimate = median,
  silent = TRUE,
  newton_config = NULL,
  useNonLinearityCorrection = FALSE,
  ...
)
}
\arguments{
\item{fit}{}

\item{fscale}{}

\item{catchval}{}

\item{fval}{}

\item{nextssb}{}

\item{landval}{}

\item{nosim}{}

\item{year.base}{}

\item{ave.years}{}

\item{rec.years}{}

\item{label}{}

\item{overwriteSelYears}{}

\item{deterministicF}{}

\item{processNoiseF}{}

\item{customSel}{}

\item{lagR}{}

\item{splitLD}{}

\item{addTSB}{}

\item{biasCorrect}{}

\item{returnAllYears}{}
}
\value{
msamforecast
}
\description{
Forecast of a multiStockassessment
}
