% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{modelforecast.msam}
\alias{modelforecast.msam}
\title{Forecast of a multiStockassessment}
\usage{
modelforecast.msam(fit, fscale = NULL, catchval = NULL, fval = NULL,
  nextssb = NULL, landval = NULL, findMSY = NULL, nosim = NULL,
  year.base = lapply(fit, function(x) max(x$data$years)),
  ave.years = lapply(fit, function(x) max(x$data$years) + (-4:0)),
  rec.years = lapply(fit, function(x) max(x$data$years) + (-9:0)),
  label = NULL, overwriteSelYears = NULL, deterministic = FALSE,
  processNoiseF = TRUE, customSel = NULL, lagR = FALSE,
  splitLD = FALSE, addTSB = FALSE, biasCorrect = TRUE,
  returnAllYears = FALSE, useUniroot = FALSE,
  nCatchAverageYears = rep(1, length(fit)), ...)
}
\arguments{
\item{fit}{}

\item{fscale}{}

\item{catchval}{}

\item{fval}{}

\item{nextssb}{}

\item{landval}{}

\item{nosim}{}

\item{year.base}{}

\item{ave.years}{}

\item{rec.years}{}

\item{label}{}

\item{overwriteSelYears}{}

\item{deterministic}{}

\item{processNoiseF}{}

\item{customSel}{}

\item{lagR}{}

\item{splitLD}{}

\item{addTSB}{}

\item{biasCorrect}{}

\item{returnAllYears}{}
}
\value{
msamforecast
}
\description{
Forecast of a multiStockassessment
}
