% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelforecast.R
\name{modelforecast.msam}
\alias{modelforecast.msam}
\title{Forecast of a multiStockassessment}
\usage{
\method{modelforecast}{msam}(
  fit,
  constraints = NULL,
  fscale = NULL,
  catchval = NULL,
  fval = NULL,
  nextssb = NULL,
  landval = NULL,
  nosim = 1000,
  year.base = unlist(min(sapply(fit, function(x) max(x$data$years)))),
  ave.years = lapply(fit, function(x) max(x$data$years) + (-9:0)),
  overwriteBioModel = FALSE,
  rec.years = lapply(fit, function(x) numeric(0)),
  label = NULL,
  overwriteSelYears = NULL,
  deterministicF = FALSE,
  processNoiseF = FALSE,
  fixedFdeviation = FALSE,
  useFHessian = FALSE,
  resampleFirst = !is.null(nosim) && nosim > 0,
  resampleParameters = FALSE,
  useModelLastN = TRUE,
  fixFirstN = FALSE,
  customSel = NULL,
  lagR = FALSE,
  splitLD = FALSE,
  addTSB = FALSE,
  biasCorrect = FALSE,
  returnAllYears = FALSE,
  returnObj = FALSE,
  progress = nosim > 0 && ncores == 1,
  estimate = median,
  silent = TRUE,
  newton_config = NULL,
  custom_pl = NULL,
  useNonLinearityCorrection = (nosim > 0 && !deterministicF),
  ncores = 1,
  ...
)
}
\arguments{
\item{fit}{}

\item{fscale}{}

\item{catchval}{}

\item{fval}{}

\item{nextssb}{}

\item{landval}{}

\item{nosim}{}

\item{year.base}{}

\item{ave.years}{}

\item{rec.years}{}

\item{label}{}

\item{overwriteSelYears}{}

\item{deterministicF}{}

\item{processNoiseF}{}

\item{customSel}{}

\item{lagR}{}

\item{splitLD}{}

\item{addTSB}{}

\item{biasCorrect}{}

\item{returnAllYears}{}
}
\value{
msamforecast
}
\description{
Forecast of a multiStockassessment
}
