% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{MSE}
\alias{MSE}
\title{Management strategy evaluation using SAM models}
\usage{
MSE(
  OM,
  EM,
  nYears,
  AdviceForecastSettings,
  AdviceYears = 1,
  AdviceLag = 1,
  initialAdvice = NA,
  implementationError = function(x) x + 1e-04,
  knotRange = 3,
  intermediateFleets = numeric(0),
  OMselectivityFixed = FALSE,
  managementToPopulation = function(x, OM) {
     splitCatch(x, OM, "KeepF", 1)
 },
  adviceToManagement = function(advice, adviceLast) advice,
  adviceMethod = c("Basic", "ICES", "Empirical"),
  EMReferencePoints = NULL,
  maxTrueF = 3,
  maxScaleF = 1.5,
  inherentImplementationError = FALSE,
  ...
)
}
\arguments{
\item{OM}{sam.fit that will work as operating model}

\item{EM}{sam.fit that will work as estimation model}

\item{nYears}{Number of years to run simulation}

\item{AdviceForecastSettings}{Settings to do forecast that determines advice}

\item{AdviceYears}{Number of years advice given at a time. How advice is given is determined by AdviceForecastSettings}

\item{AdviceLag}{Lag between assessment and advice}

\item{initialAdvice}{Advice in the first AdviceLag years}

\item{implementationError}{Function to add implementation error (i.e, transform advice to target catch)}

\item{knotRange}{Range of spline knot values to try}

\item{intermediateFleets}{Fleets that are available in the (first) intermediate year}

\item{OMselectivityFixed}{Fix selectivity in OM?}

\item{...}{arguments passed on to addSimulatedYears}
}
\value{
a list with MSE result
}
\description{
Management strategy evaluation using SAM models
}
