% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisam.R
\name{multisam.fit}
\alias{multisam.fit}
\title{Fit multiple SAM models with correlations}
\usage{
multisam.fit(
  x,
  formula = ~-1,
  corStructure = suggestCorStructure(x, nAgeClose = 0),
  usePartialCors = TRUE,
  newtonsteps = 0,
  rm.unidentified = FALSE,
  nlminb.control = list(trace = 1, eval.max = 20000, iter.max = 20000),
  lower = NULL,
  upper = NULL,
  starting = NULL,
  community_formula = ~-1,
  community_type = 1,
  shared_data = NULL,
  shared_keys = character(0),
  shared_selectivity = 0,
  shared_seasonality = 0,
  shared_stockrecruitment = FALSE,
  shared_oneFScalePars = FALSE,
  shared_initN = FALSE,
  shared_fleetParameters = c(),
  shared_proportionalHazard = NULL,
  shared_phmap = NULL,
  skip_stock_observations = FALSE,
  stockAreas = matrix(1, 1, length(x)),
  genetics_data = prepareGenetics(),
  genetics_dirichlet = FALSE,
  genetics_spatialAge = TRUE,
  genetics_independentStocks = TRUE,
  initN = 0,
  initF = FALSE,
  parlist = NULL,
  run = TRUE,
  symbolicAnalysis = FALSE,
  fullDerived = FALSE,
  mohn = FALSE,
  doSdreport = TRUE,
  skip.hessian = FALSE,
  ...
)
}
\arguments{
\item{x}{samset from the stockassessment package}

\item{formula}{formula for covariance matrix covariates (See Details)}

\item{corStructure}{symmetric boolean matrix. True if a (partial) correlation in survival should be fixed to zero between the corresponding age/stock combination}

\item{usePartialCors}{if TRUE corStructure describes the partial correlations. If FALSE corStructure describes correlations.}

\item{newtonsteps}{As for stockassessment::sam.fit}

\item{lower}{As for stockassessment::sam.fit}

\item{upper}{As for stockassessment::sam.fit}

\item{...}{Additional arguments passed to TMB::MakeADFun}
}
\value{
A list of class msam and samset, which is the input x with additional information attached.
}
\description{
Fit multiple SAM models with correlated survival processes.
}
\details{
Function to fit a multi-stock SAM model (Albertsen et al., 2018).
}
\examples{
\donttest{
if(require(stockassessment)){
  data(nscodData)
  data(nscodConf)
  data(nscodParameters)
  fit <- sam.fit(nscodData, nscodConf, nscodParameters)
  fits <- c(fit)
  obj <- multisam.fit(fits)
}
}
}
\references{
Albertsen, C. M., Nielsen, A. and Thygesen, U. H. (2018) Connecting single-stock assessment models through correlated survival. ICES Journal of Marine Science, 75(1), 235-244. doi: 10.1093/icesjms/fsx114
}
\author{
Christoffer Moesgaard Albertsen
}
