% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisam.R
\name{multisam.fit}
\alias{multisam.fit}
\title{Fit multiple SAM models with correlations}
\usage{
multisam.fit(x, corStructure, usePartialCors = TRUE, newtonsteps = 3,
  lower = NULL, upper = NULL, ...)
}
\arguments{
\item{x}{samset from the stockassessment package}

\item{corStructure}{symmetric boolean matrix. True if a (partial) correlation in survival should be estimated between the corresponding age/stock combination}

\item{usePartialCors}{if TRUE corStructure describes the partial correlations. If FALSE corStructure describes correlations.}

\item{newtonsteps}{As for stockassessment::sam.fit}

\item{lower}{As for stockassessment::sam.fit}

\item{upper}{As for stockassessment::sam.fit}

\item{...}{Additional arguments passed to TMB::MakeADFun}
}
\value{
A list of class msam and samset
}
\description{
Fit multiple SAM models with correlated survival processes.
}
\examples{
\donttest{
if(require(stockassessment)){
  data(nscodData)
  data(nscodConf)
  data(nscodParameters)
  fit <- sam.fit(nscodData, nscodConf, nscodParameters)
  fits <- c(fit)
  obj <- multisam.fit(fits)
}
}
}
\author{
Christoffer Moesgaard Albertsen
}
