% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotit.msam}
\alias{plotit.msam}
\alias{plotit.msamforecast}
\title{Function to actually do the plotting}
\usage{
\method{plotit}{msam}(
  fit,
  what,
  x = lapply(attr(fit, "m_data")$sam, function(x) x$years),
  ylab = what,
  xlab = "Years",
  ex = numeric(0),
  trans = function(x) x,
  add = FALSE,
  ci = TRUE,
  cicol = gray(0.5, alpha = 0.5),
  addCI = NA,
  drop = 0,
  unnamed.basename = "current",
  xlim = NULL,
  ciAlpha = 0.3,
  col = .plotcols.crp(length(fit)),
  extraLabel = NULL,
  addTotal = FALSE,
  onlyTotal = FALSE,
  legend.pos = "bottom",
  ...
)

\method{plotit}{msamforecast}(
  fit,
  what,
  x = fit$data$years,
  ylab = what,
  xlab = "Years",
  ex = numeric(0),
  trans = exp,
  add = FALSE,
  ci = TRUE,
  cicol = gray(0.5, alpha = 0.5),
  addCI = NA,
  drop = 0,
  unnamed.basename = "current",
  xlim = NULL,
  addTotal = FALSE,
  onlyTotal = FALSE,
  ...
)
}
\arguments{
\item{fit}{msam fit}

\item{what}{variable to plot}

\item{x}{x-axis values}

\item{ylab}{y-axis label}

\item{xlab}{x-axis label}

\item{ex}{extra years to add}

\item{trans}{function to transform variable to plot}

\item{add}{If false a new plot is created. If true everything is added to the previous plot-}

\item{ci}{Add confidence intervals?}

\item{cicol}{Color of confidence intervals}

\item{addCI}{Not used}

\item{drop}{Number of years to drop from the end}

\item{unnamed.basename}{not used}

\item{xlim}{x-axis limits. If null, the range of x is used.}

\item{ciAlpha}{Alpha channel value of confidence interval color}

\item{col}{line colors}

\item{extraLabel}{Not used}

\item{...}{Other arguments}
}
\description{
Function to actually do the plotting
}
