% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{srplot.msam}
\alias{srplot.msam}
\title{Stock-recruitment plot}
\usage{
\method{srplot}{msam}(
  fit,
  textcol = "red",
  add = FALSE,
  col = .plotcols.crp(length(fit)),
  ...
)
}
\arguments{
\item{fit}{msam object}

\item{textcol}{Text color}

\item{add}{Should the figure be added to a current plot?}

\item{col}{Line colors}

\item{...}{extra arguments for plotting}
}
\description{
Stock-recruitment plot
}
\author{
Christoffer Moesgaard Albertsen
}
